+++ scarletio/utils/trace/frame_proxy/tests/test__frame_proxying.py
@@ -92,3 +92,32 @@ def test__populate_frame_proxies():
         vampytest.assert_instance(frame_proxy.expression_info, ExpressionInfo)
     
     vampytest.assert_eq(frame_proxy.mention_count, 2)
+
+
+def test__populate_frame_proxies__desync():
+    """
+    Tests whether ``populate_frame_proxies`` works as intended.
+    
+    Case: Desync while populating.
+    """
+    frame_proxy = FrameProxyVirtual.from_fields(file_name = 'koishi.py')
+    expression_info = ExpressionInfo(frame_proxy.expression_key, [], 0, True)
+    new_file_name = 'kokoro.py'
+    
+    def mock_get_expression_info(expression_key):
+        nonlocal frame_proxy
+        nonlocal expression_info
+        nonlocal new_file_name
+        vampytest.assert_eq(expression_key, expression_info.key)
+        frame_proxy.file_name = new_file_name
+        return expression_info
+    
+    frame_proxies = [frame_proxy]
+    
+    mocked = vampytest.mock_globals(populate_frame_proxies, get_expression_info = mock_get_expression_info)
+    mocked(frame_proxies)
+    
+    for frame_proxy in frame_proxies:
+        vampytest.assert_instance(frame_proxy.expression_info, ExpressionInfo)
+    
+    vampytest.assert_eq(frame_proxy.expression_info.key.file_name, new_file_name)